pyproject.toml
[project]
name = "credit_risk_model"
version = "0.0.1"
description = "This is a template repository for Python projects that use uv for their dependency management."
authors = [{ name = "Abhishek kumar", email = "abhishekbjaiswar@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11"

dependencies = [
    "boto3>=1.42.25",
    "dvc[s3]>=2.1.0",
    "fastapi>=0.128.0",
    "gunicorn>=23.0.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.6",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "uvicorn>=0.40.0",
    "xgboost>=3.1.3",
]

[project.urls]
Homepage = "https://@jaisinfinity.github.io/example-project/"
Repository = "https://github.com/@jaisinfinity/example-project"
Documentation = "https://@jaisinfinity.github.io/example-project/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",
    
    "mypy>=1.19.1",
    
    "ruff>=0.14.10",
    "httpx>=0.28.1",
    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/credit_risk_model"]

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]  # at the test time only test from tests file. just run "pytest"
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 80
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true


